make Rational to hold BigDecimal

+++ PealResultVerifier/src/main/scala/peal/verifier/util/Rational.scala
@@ -2,9 +2,9 @@ package peal.verifier.util
 
 object Rational {
   def apply(num: String) : Rational = Rational(num, "1")
-  def apply(enum: String, denom: String) = new Rational(enum.toDouble.toInt, denom.toDouble.toInt)
+  def apply(enum: String, denom: String) = new Rational(enum.toDouble, denom.toDouble)
 }
 
-case class Rational(numerator: Int, denominator: Int) {
-  def value : BigDecimal = numerator.toDouble / denominator
+case class Rational(numerator: BigDecimal, denominator: BigDecimal) {
+  def value : BigDecimal = numerator / denominator
 }
